// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        User: {
            id: components["schemas"]["Uuid"];
            /** Format: email */
            email: string;
            display_name: string;
            role: components["schemas"]["Role"];
            created_at: components["schemas"]["Timestamp"];
            /** Format: date-time */
            disabled_at?: string | null;
        };
// ...
    };
// ...
}